optimization base filter for buffer&state allocating for MetalImage

+++ MetalImage/MetalImage/Shaders/imgExposure.metal
@@ -10,13 +10,13 @@
 #include <metal_stdlib>
 using namespace metal;
 kernel void exposure(
-                     texture2d<float, access::read> inTexture [[texture(0)]],
-                     texture2d<float, access::write> outTexture [[texture(1)]],
+                     texture2d<half, access::read> inTexture [[texture(0)]],
+                     texture2d<half, access::write> outTexture [[texture(1)]],
                      device float *exposure [[buffer(0)]],
                      uint2 gid [[thread_position_in_grid]])
 {
-    const float4 inColor = inTexture.read(gid);
-    const float4 outColor(inColor.rgb * float3(pow(2.0, *exposure)),inColor.a);
+    const half4 inColor = inTexture.read(gid);
+    const half4 outColor(inColor.rgb * half3(pow(2.0, *exposure)),inColor.a);
     outTexture.write(outColor, gid);
 }
